feat : Comment User 매핑

+++ src/main/java/techcourse/fakebook/service/CommentService.java
@@ -4,9 +4,12 @@ import org.springframework.stereotype.Service;
 import techcourse.fakebook.domain.article.Article;
 import techcourse.fakebook.domain.comment.Comment;
 import techcourse.fakebook.domain.comment.CommentRepository;
+import techcourse.fakebook.domain.user.User;
+import techcourse.fakebook.exception.InvalidAuthorException;
 import techcourse.fakebook.exception.NotFoundCommentException;
 import techcourse.fakebook.service.dto.CommentRequest;
 import techcourse.fakebook.service.dto.CommentResponse;
+import techcourse.fakebook.service.dto.UserDto;
 import techcourse.fakebook.utils.CommentAssembler;
 
 import javax.transaction.Transactional;
@@ -17,10 +20,12 @@ import java.util.stream.Collectors;
 @Transactional
 public class CommentService {
     private ArticleService articleService;
+    private UserService userService;
     private CommentRepository commentRepository;
 
-    public CommentService(ArticleService articleService, CommentRepository commentRepository) {
+    public CommentService(ArticleService articleService, UserService userService, CommentRepository commentRepository) {
         this.articleService = articleService;
+        this.userService = userService;
         this.commentRepository = commentRepository;
     }
 
@@ -35,20 +40,23 @@ public class CommentService {
                 .collect(Collectors.toList());
     }
 
-    public CommentResponse save(Long articleId, CommentRequest commentRequest) {
+    public CommentResponse save(Long articleId, CommentRequest commentRequest, UserDto userDto) {
+        User user = userService.getUser(userDto.getId());
         Article article = articleService.getArticle(articleId);
-        Comment comment = commentRepository.save(CommentAssembler.toEntity(commentRequest, article));
+        Comment comment = commentRepository.save(CommentAssembler.toEntity(commentRequest, article, user));
         return CommentAssembler.toResponse(comment);
     }
 
-    public CommentResponse update(Long id, CommentRequest updatedRequest) {
+    public CommentResponse update(Long id, CommentRequest updatedRequest, UserDto userDto) {
         Comment comment = getComment(id);
+        checkAuthor(userDto, comment);
         comment.update(updatedRequest.getContent());
         return CommentAssembler.toResponse(comment);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, UserDto userDto) {
         Comment comment = getComment(id);
+        checkAuthor(userDto, comment);
         comment.delete();
     }
 
@@ -58,5 +66,11 @@ public class CommentService {
             throw new NotFoundCommentException();
         }
         return comment;
+    }
+
+    private void checkAuthor(UserDto userDto, Comment comment) {
+        if (comment.getUser().isNotAuthor(userDto.getId())) {
+            throw new InvalidAuthorException();
+        }
     }
 }
